Split weed.py into two master.py and volume.py
