+++ weed/__init__.py
@@ -1,7 +1,7 @@
-import weed
-from weed import *
+from master import WeedMaster
+from volume import WeedVolume
 from .version import VERSION, __version__
 
-__all__ = weed.__all__
+__all__ = ['WeedMaster', 'WeedVolume']
 
 __author__ = 'darkdarkfruit'
